+++ packages/presets/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `DatePicker` component.
+
 ## [0.4.0] - 2023-08-10
 
 ### Added